+++ composeApp/src/commonMain/kotlin/org/wip/plugintoolkit/ui/AppScaffold.kt
@@ -53,7 +53,6 @@ fun AppScaffold(
                     content()
                 }
             }
-
             NavigationSidebar(
                 title = Res.string.app_name.localized,
                 bodySections = sections,
@@ -65,10 +64,10 @@ fun AppScaffold(
                 modifier = Modifier.fillMaxHeight()
             )
 
-            ToastHost(
-                notificationService = notificationService, settings = settings.notifications
-            )
-            DialogHost(dialogService)
+            // ToastHost(
+            //     notificationService = notificationService, settings = settings.notifications
+            // )
+            // DialogHost(dialogService)
         }
     }
 }