+++ composeApp/src/commonTest/kotlin/org/wip/plugintoolkit/features/flows/JobWorkerFlowBasicTest.kt
@@ -58,7 +58,7 @@ class JobWorkerFlowBasicTest : JobWorkerFlowTestBase() {
 
         val outputs = org.wip.plugintoolkit.features.job.logic.FlowEngine(
             jobManager,
-            org.wip.plugintoolkit.features.job.logic.DefaultSystemNodeExecutorRegistry(),
+            org.wip.plugintoolkit.features.job.logic.DefaultSystemNodeExecutorRegistry(mockk(relaxed = true)),
             mockk(relaxed = true),
             mockk(relaxed = true),
             backgroundScope
@@ -104,7 +104,7 @@ class JobWorkerFlowBasicTest : JobWorkerFlowTestBase() {
         // Should complete without throwing exception (soft failure)
         val outputs = org.wip.plugintoolkit.features.job.logic.FlowEngine(
             jobManager,
-            org.wip.plugintoolkit.features.job.logic.DefaultSystemNodeExecutorRegistry(),
+            org.wip.plugintoolkit.features.job.logic.DefaultSystemNodeExecutorRegistry(mockk(relaxed = true)),
             mockk(relaxed = true),
             mockk(relaxed = true),
             backgroundScope
@@ -147,7 +147,7 @@ class JobWorkerFlowBasicTest : JobWorkerFlowTestBase() {
         try {
             org.wip.plugintoolkit.features.job.logic.FlowEngine(
                 jobManager,
-                org.wip.plugintoolkit.features.job.logic.DefaultSystemNodeExecutorRegistry(),
+                org.wip.plugintoolkit.features.job.logic.DefaultSystemNodeExecutorRegistry(mockk(relaxed = true)),
                 mockk(relaxed = true),
                 mockk(relaxed = true),
                 backgroundScope