+++ composeApp/src/commonTest/kotlin/org/wip/plugintoolkit/features/plugin/logic/PluginLifecycleManagerTest.kt
@@ -37,8 +37,8 @@ class PluginLifecycleManagerTest {
 
     private class FakeSettingsPersistence : SettingsPersistence {
         var settings = AppSettings()
-        override fun load(): AppSettings = settings
-        override fun save(settings: AppSettings) {
+        override suspend fun load(): AppSettings = settings
+        override suspend fun save(settings: AppSettings) {
             this.settings = settings
         }
 
@@ -53,7 +53,8 @@ class PluginLifecycleManagerTest {
         val fileSystem = FakeFileSystem()
         val persistence = FakeSettingsPersistence()
         val settingsRepo = SettingsRepository(persistence, backgroundScope)
-        val registry = PluginRegistry(settingsRepo, backgroundScope, loomDispatcher)
+        val mockAppConfig = io.mockk.mockk<org.wip.plugintoolkit.core.SystemConfig>(relaxed = true)
+        val registry = PluginRegistry(settingsRepo, backgroundScope, loomDispatcher, mockAppConfig)
         val jobManager = JobManager(backgroundScope, settingsRepo)
         val lifecycleManager = PluginLifecycleManager(registry, jobManager, settingsRepo, fileSystem)
 
@@ -100,7 +101,8 @@ class PluginLifecycleManagerTest {
         val fileSystem = FakeFileSystem()
         val persistence = FakeSettingsPersistence()
         val settingsRepo = SettingsRepository(persistence, backgroundScope)
-        val registry = PluginRegistry(settingsRepo, backgroundScope, loomDispatcher)
+        val mockAppConfig = io.mockk.mockk<org.wip.plugintoolkit.core.SystemConfig>(relaxed = true)
+        val registry = PluginRegistry(settingsRepo, backgroundScope, loomDispatcher, mockAppConfig)
         val jobManager = JobManager(backgroundScope, settingsRepo)
         val lifecycleManager = PluginLifecycleManager(registry, jobManager, settingsRepo, fileSystem)